Pick some low-hanging warning fruit by changing several integer variables to size_t or apr_size_t, depending on the context of usage.  This removes a number of "implicit conversion shortens 64-bit value into a 32-bit value" warnings.
* subversion/svn/util.c
  (svn_cl__get_log_message),
* subversion/libsvn_client/merge.c
  (fix_deleted_subtree_ranges, populate_remaining_ranges,
   record_mergeinfo_for_dir_merge),
* subversion/libsvn_client/locking_commands.c
  (organize_lock_targets),
* subversion/libsvn_client/commit_util.c
  (svn_client__condense_commit_items),
* subversion/libsvn_client/add.c
  (trim_string, auto_props_enumerator):
  Update variable types.

* subversion/svn/main.c
  (main):
  Cast value to (int).


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@879075 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/commit_util.c
@@ -1252,8 +1252,8 @@ svn_client__condense_commit_items(const char **base_url,
     {
       svn_client_commit_item3_t *this_item
         = APR_ARRAY_IDX(ci, i, svn_client_commit_item3_t *);
-      int url_len = strlen(this_item->url);
-      int base_url_len = strlen(*base_url);
+      size_t url_len = strlen(this_item->url);
+      size_t base_url_len = strlen(*base_url);
 
       if (url_len > base_url_len)
         this_item->url = apr_pstrdup(pool, this_item->url + base_url_len + 1);
